// src/app/openapi/model/fastPost.ts
/**
 * SpeedyGo API
 * Navigators Team API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FastPost { 
    idF?: string;
    receiverName?: string;
    receiverAddress?: string;
    receiverTelNbr?: number;
    packageWeight?: string;
    fastPostStatus?: FastPost.FastPostStatusEnum;
}
export namespace FastPost {
    export type FastPostStatusEnum = 'PENDING' | 'APPROVED' | 'REJECTED';
    export const FastPostStatusEnum = {
        Pending: 'PENDING' as FastPostStatusEnum,
        Approved: 'APPROVED' as FastPostStatusEnum,
        Rejected: 'REJECTED' as FastPostStatusEnum
    };
}